pyproject.toml
[project]
name = "green-agent-template"
version = "0.1.0"
description = "A template for A2A green agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "mlebench",
    "numpy>=2.0.0",
    "openai>=2.0.0",
    "pandas>=2.2.0",
    "pydantic>=2.12.5",
    "scikit-learn>=1.5.0",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]

[tool.uv.sources]
mlebench = { git = "https://github.com/LineaLabs/mle-bench.git", rev = "b43c56d8dd0ae5be74353597ab42241f0cb9a997" }
